pcd: migrate k(Y) evaluation from hashes_1 to outer_collapse for gate budget
